Move column leaves to common-storage
