+++ src/common/storage/src/lib.rs
@@ -68,3 +68,7 @@ mod utils;
 
 mod cache;
 pub use cache::FuseCachePolicy;
+
+mod column_leaf;
+pub use column_leaf::ColumnLeaf;
+pub use column_leaf::ColumnLeaves;
