+++ src/query/catalog/src/plan/projection.rs
@@ -16,6 +16,9 @@ use std::collections::BTreeMap;
 use std::fmt::Formatter;
 
 use common_datavalues::DataSchema;
+use common_exception::Result;
+use common_storage::ColumnLeaf;
+use common_storage::ColumnLeaves;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, PartialEq, Eq)]
 pub enum Projection {
@@ -48,6 +51,28 @@ impl Projection {
             Projection::Columns(indices) => schema.project(indices),
             Projection::InnerColumns(path_indices) => schema.inner_project(path_indices),
         }
+    }
+
+    pub fn project_column_leaves<'a>(
+        &'a self,
+        column_leaves: &'a ColumnLeaves,
+    ) -> Result<Vec<&ColumnLeaf>> {
+        let column_leaves = match self {
+            Projection::Columns(indices) => indices
+                .iter()
+                .map(|idx| &column_leaves.column_leaves[*idx])
+                .collect(),
+            Projection::InnerColumns(path_indices) => {
+                let paths: Vec<&Vec<usize>> = path_indices.values().collect();
+                paths
+                    .iter()
+                    .map(|path| {
+                        ColumnLeaves::traverse_path(&column_leaves.column_leaves, path).unwrap()
+                    })
+                    .collect()
+            }
+        };
+        Ok(column_leaves)
     }
 }
