+++ src/query/storages/fuse/fuse/src/fuse_part.rs
@@ -19,12 +19,8 @@ use std::hash::Hash;
 use std::hash::Hasher;
 use std::sync::Arc;
 
-use common_arrow::arrow::datatypes::DataType as ArrowType;
-use common_arrow::arrow::datatypes::Field as ArrowField;
-use common_arrow::arrow::datatypes::Schema as ArrowSchema;
 use common_catalog::plan::PartInfo;
 use common_catalog::plan::PartInfoPtr;
-use common_catalog::plan::Projection;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_storages_table_meta::meta::Compression;
@@ -45,7 +41,6 @@ impl ColumnMeta {
         }
     }
 }
-
 #[derive(serde::Serialize, serde::Deserialize, PartialEq, Eq)]
 pub struct FusePartInfo {
     pub location: String,
@@ -100,101 +95,6 @@ impl FusePartInfo {
             None => Err(ErrorCode::Internal(
                 "Cannot downcast from PartInfo to FusePartInfo.",
             )),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct ColumnLeaves {
-    pub column_leaves: Vec<ColumnLeaf>,
-}
-
-impl ColumnLeaves {
-    pub fn new_from_schema(schema: &ArrowSchema) -> Self {
-        let mut leaf_id = 0;
-        let mut column_leaves = Vec::with_capacity(schema.fields.len());
-
-        for field in &schema.fields {
-            let column_leaf = Self::traverse_fields_dfs(field, &mut leaf_id);
-            column_leaves.push(column_leaf);
-        }
-
-        Self { column_leaves }
-    }
-
-    fn traverse_fields_dfs(field: &ArrowField, leaf_id: &mut usize) -> ColumnLeaf {
-        match &field.data_type {
-            ArrowType::Struct(inner_fields) => {
-                let mut child_column_leaves = Vec::with_capacity(inner_fields.len());
-                let mut child_leaf_ids = Vec::with_capacity(inner_fields.len());
-                for inner_field in inner_fields {
-                    let child_column_leaf = Self::traverse_fields_dfs(inner_field, leaf_id);
-                    child_leaf_ids.extend(child_column_leaf.leaf_ids.clone());
-                    child_column_leaves.push(child_column_leaf);
-                }
-                ColumnLeaf::new(field.clone(), child_leaf_ids, Some(child_column_leaves))
-            }
-            _ => {
-                let column_leaf = ColumnLeaf::new(field.clone(), vec![*leaf_id], None);
-                *leaf_id += 1;
-                column_leaf
-            }
-        }
-    }
-
-    pub fn get_by_projection<'a>(&'a self, proj: &'a Projection) -> Result<Vec<&ColumnLeaf>> {
-        let column_leaves = match proj {
-            Projection::Columns(indices) => indices
-                .iter()
-                .map(|idx| &self.column_leaves[*idx])
-                .collect(),
-            Projection::InnerColumns(path_indices) => {
-                let paths: Vec<&Vec<usize>> = path_indices.values().collect();
-                paths
-                    .iter()
-                    .map(|path| Self::traverse_path(&self.column_leaves, path).unwrap())
-                    .collect()
-            }
-        };
-        Ok(column_leaves)
-    }
-
-    fn traverse_path<'a>(
-        column_leaves: &'a [ColumnLeaf],
-        path: &'a [usize],
-    ) -> Result<&'a ColumnLeaf> {
-        let column_leaf = &column_leaves[path[0]];
-        if path.len() > 1 {
-            return match &column_leaf.children {
-                Some(ref children) => Self::traverse_path(children, &path[1..]),
-                None => Err(ErrorCode::Internal(format!(
-                    "Cannot get column_leaf by path: {:?}",
-                    path
-                ))),
-            };
-        }
-        Ok(column_leaf)
-    }
-}
-
-/// `ColumnLeaf` contains all the leaf column ids of the column.
-/// For the nested types, it may contain more than one leaf column.
-#[derive(Debug, Clone)]
-pub struct ColumnLeaf {
-    pub field: ArrowField,
-    // `leaf_ids` is the indices of all the leaf columns in DFS order,
-    // through which we can find the meta information of the leaf columns.
-    pub leaf_ids: Vec<usize>,
-    // Optional children column for nested types.
-    pub children: Option<Vec<ColumnLeaf>>,
-}
-
-impl ColumnLeaf {
-    pub fn new(field: ArrowField, leaf_ids: Vec<usize>, children: Option<Vec<ColumnLeaf>>) -> Self {
-        Self {
-            field,
-            leaf_ids,
-            children,
         }
     }
 }
