+++ src/query/storages/fuse/fuse/src/lib.rs
@@ -35,8 +35,6 @@ use common_catalog::table::TableStatistics;
 pub use common_catalog::table_context::TableContext;
 use common_catalog::table_mutator::TableMutator;
 pub use constants::*;
-pub use fuse_part::ColumnLeaf;
-pub use fuse_part::ColumnLeaves;
 pub use fuse_part::ColumnMeta;
 pub use fuse_part::FusePartInfo;
 pub use fuse_table::FuseTable;
